Fixed some bugs

+++ server/tasks/structs.go
@@ -220,7 +220,7 @@ type DeezerTrack struct {
 	ExplicitContentLyrics int      `json:"explicit_content_lyrics"`
 	ExplicitContentCover  int      `json:"explicit_content_cover"`
 	Preview               string   `json:"preview"`
-	Bpm                   int      `json:"bpm"`
+	Bpm                   float64  `json:"bpm"`
 	Gain                  float64  `json:"gain"`
 	AvailableCountries    []string `json:"available_countries"`
 	Contributors          []struct {
